Enabled package to deal with Decimals
Before, it was converting values to floats raising an exception if the
conversion lost precision when being cast to int.  Throwing an exception
stored a pickled value, and before Decimals would be converted to floats
before raising this exception and being pickled.  Now, it only converts
to float if the value is not a decimal instance.  Tested it
in shell, and it works!